More linting fixes, added travis.yml
