+++ .travis.yml
@@ -1,0 +1,9 @@
+language: node_js
+node_js:
+  - "node"
+script:
+  - npm i
+  - npm run lint
+branches:
+  only:
+    - main
