+++ components/CreateWorkout/CreateWorkout.jsx
@@ -1,7 +1,6 @@
-/* eslint-disable react/prop-types */
 import { AntDesign } from '@expo/vector-icons';
 import * as React from 'react';
-import { Alert, View } from 'react-native';
+import { View } from 'react-native';
 import styled from 'styled-components/native';
 import { useDispatch, useSelector } from 'react-redux';
 import * as firebase from 'firebase';
@@ -123,30 +122,30 @@ export default ({ navigation }) => {
         } else if (itemState.length === 0) {
           alert('Please add at least one exercise');
         } else {
-          let muscleGroups = itemState[0].muscleGroups; 
+          let { muscleGroups } = itemState[0];
           if (itemState.length >= 2) { // for showing top 2 muscle groups
-            muscleGroups += " ";
+            muscleGroups += ' ';
             muscleGroups += itemState[1].muscleGroups;
           }
           const newWorkout = {
             name,
             lastPerformed: 'n/a',
             id: nextWorkoutId,
-            muscleGroups: muscleGroups,
+            muscleGroups,
             color: itemState[0].color,
             exercises: itemState.map((item) => {
               const setArr = [];
               const sets = item.sets || 3;
-              for (let i = 0; i < sets; i++) {
-                item.isReps ?
-                  setArr.push({ weight: undefined, reps: item.reps || 10 }) :
-                  setArr.push({ weight: undefined, time: item.seconds || 60 })
+              for (let i = 0; i < sets; i += 1) {
+                item.isReps
+                  ? setArr.push({ weight: undefined, reps: item.reps || 10 })
+                  : setArr.push({ weight: undefined, time: item.seconds || 60 });
               }
               return {
                 sets: setArr,
                 exerciseId: item.id,
               };
-            })
+            }),
           };
           sendWorkoutToDB(newWorkout);
           dispatch(actions.workouts.addWorkout(newWorkout));
